Rename pkg name to husarion_mecanum_drive_controller
