+++ husarion_mecanum_drive_controller/CMakeLists.txt
@@ -2,7 +2,7 @@
 # (https://github.com/ros-controls/ros2_controllers)
 
 cmake_minimum_required(VERSION 3.11)
-project(mecanum_drive_controller)
+project(husarion_mecanum_drive_controller)
 
 # Default to C++17
 if(NOT CMAKE_CXX_STANDARD)
@@ -38,33 +38,37 @@ endforeach()
 get_target_property(TB_INCLUDE_DIRS control_toolbox::rate_limiter_parameters
                     INTERFACE_INCLUDE_DIRECTORIES)
 generate_parameter_library(
-  mecanum_drive_controller_parameters
-  src/mecanum_drive_controller_parameter.yaml
+  husarion_mecanum_drive_controller_parameters
+  src/husarion_mecanum_drive_controller_parameter.yaml
   ${TB_INCLUDE_DIRS}/control_toolbox/custom_validators.hpp)
 
-add_library(mecanum_drive_controller SHARED src/mecanum_drive_controller.cpp
-                                            src/odometry.cpp)
+add_library(husarion_mecanum_drive_controller SHARED
+            src/husarion_mecanum_drive_controller.cpp src/odometry.cpp)
 target_include_directories(
-  mecanum_drive_controller
+  husarion_mecanum_drive_controller
   PUBLIC $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
-         $<INSTALL_INTERFACE:include/mecanum_drive_controller>)
+         $<INSTALL_INTERFACE:include/husarion_mecanum_drive_controller>)
 target_link_libraries(
-  mecanum_drive_controller PUBLIC mecanum_drive_controller_parameters
-                                  control_toolbox::rate_limiter_parameters)
-ament_target_dependencies(mecanum_drive_controller PUBLIC
+  husarion_mecanum_drive_controller
+  PUBLIC husarion_mecanum_drive_controller_parameters
+         control_toolbox::rate_limiter_parameters)
+ament_target_dependencies(husarion_mecanum_drive_controller PUBLIC
                           ${THIS_PACKAGE_INCLUDE_DEPENDS})
 pluginlib_export_plugin_description_file(controller_interface
                                          mecanum_drive_plugin.xml)
 
-install(DIRECTORY include/ DESTINATION include/mecanum_drive_controller)
+install(DIRECTORY include/
+        DESTINATION include/husarion_mecanum_drive_controller)
 
 install(
-  TARGETS mecanum_drive_controller mecanum_drive_controller_parameters
-  EXPORT export_mecanum_drive_controller
+  TARGETS husarion_mecanum_drive_controller
+          husarion_mecanum_drive_controller_parameters
+  EXPORT export_husarion_mecanum_drive_controller
   RUNTIME DESTINATION bin
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib)
 
-ament_export_targets(export_mecanum_drive_controller HAS_LIBRARY_TARGET)
+ament_export_targets(export_husarion_mecanum_drive_controller
+                     HAS_LIBRARY_TARGET)
 ament_export_dependencies(${THIS_PACKAGE_INCLUDE_DEPENDS})
 ament_package()
