+++ husarion_mecanum_drive_controller/mecanum_drive_plugin.xml
@@ -1,6 +1,6 @@
 <!-- Copied and adapted from diff_drive_controller (https://github.com/ros-controls/ros2_controllers) -->
-<library path="mecanum_drive_controller">
-  <class name="mecanum_drive_controller/MecanumDriveController" type="mecanum_drive_controller::MecanumDriveController" base_class_type="controller_interface::ChainableControllerInterface">
+<library path="husarion_mecanum_drive_controller">
+  <class name="husarion_mecanum_drive_controller/MecanumDriveController" type="husarion_mecanum_drive_controller::MecanumDriveController" base_class_type="controller_interface::ChainableControllerInterface">
   <description>
     The mecanum drive controller transforms linear and angular velocity messages into signals for each wheel(s) for a mecanum drive robot.
   </description>
